Remove Typed Clients; Move all logic back to services; Create HttpClient singleton directly in Program.cs with AccountService.cs having its own HttpClient
