+++ DarkBot/Web/Components/Home/ChangeCurrentGuild.razor
@@ -1,7 +1,6 @@
 ﻿@using DarkBot.Models
 @using Newtonsoft.Json
 @using System.ComponentModel.DataAnnotations
-@using DarkBot.Clients
 @using DarkBot.Services
 @using Microsoft.IdentityModel.Tokens
 @using Microsoft.VisualBasic
