+++ DarkBot/Web/Components/Home/CurrentGuild.razor
@@ -1,6 +1,5 @@
 ﻿@using DarkBot.Models
 @using Newtonsoft.Json
-@using DarkBot.Clients
 @using DarkBot.Services
 <div class="card max-w-96 bg-base-200 border-2 border-neutral mb-4">
     @if (Guild is null && !BotNotSet)
