+++ DarkBot/Web/Pages/QotPage.razor.cs
@@ -1,5 +1,4 @@
-﻿using DarkBot.Clients;
-using DarkBot.Models;
+﻿using DarkBot.Models;
 using DarkBot.Services;
 using Microsoft.AspNetCore.Components;
 using Microsoft.JSInterop;
